updated projects state and context
